requirements.txt
fastapi
uvicorn
sqlalchemy
python-jose
passlib[bcrypt]
pydantic
python-dotenv
pydantic[email]